article-summarizer/requirements.txt
ipykernel
unstructured
tiktoken
pinecone-client
pypdf
openai
langchain
pandas
numpy
python-dotenv
arxiv
pymupdf
IProgress
streamlit